fix bug
这个bug在windows会报(void *)未知的大小，还没有测试linux会不会有问题

+++ storage/allocator/yac_allocator.c
@@ -108,7 +108,7 @@ do_alloc:
 		pos += size;
 		segment->pos = pos;
 		if (segment->pos >= pos) {
-			return (void *)(segment->p + (pos - size));
+			return (void *)((unsigned int)segment->p + (pos - size));
 		} else if (retry--) {
 			goto do_retry;
 		}
